chore: rename alicloud chat model to qwen cloud
