+++ packages/@n8n/nodes-langchain/package.json
@@ -119,7 +119,7 @@
       "dist/nodes/embeddings/EmbeddingsOllama/EmbeddingsOllama.node.js",
       "dist/nodes/embeddings/EmbeddingsOracleDB/EmbeddingsOracleDb.node.js",
       "dist/nodes/embeddings/EmbeddingsNvidia/EmbeddingsNvidia.node.js",
-      "dist/nodes/llms/LmChatAlibabaCloud/LmChatAlibabaCloud.node.js",
+      "dist/nodes/llms/LmChatQwenCloud/LmChatQwenCloud.node.js",
       "dist/nodes/llms/LMChatAnthropic/LmChatAnthropic.node.js",
       "dist/nodes/llms/LmChatAzureOpenAi/LmChatAzureOpenAi.node.js",
       "dist/nodes/llms/LmChatAwsBedrock/LmChatAwsBedrock.node.js",
